+++ src/worker/execution/node-executor/script/squishy/io/squishy-io.ts
@@ -1,10 +1,12 @@
+import * as xmldom from 'xmldom'
 
 export class SquishyIO {
 
     static NEWLINE: string = "\n"
     static DELIMITER: string = ';'
+    static XML_MIME_TYPE: string = 'application/xml'
 
-    csvStringify(data: (unknown[][]) | undefined, delimiter?: string, newline?: string): string {
+    csvStringify(data: (unknown[][]) | undefined, toString?: (x: unknown) => string, delimiter?: string, newline?: string): string {
         if (!data) {
             return ''
         }
@@ -17,9 +19,16 @@ export class SquishyIO {
             newline = SquishyIO.NEWLINE
         }
 
+        if (!toString) {
+            // create default to string
+            toString = (x: unknown) => {
+                return `${x}`
+            }
+        }
+
         return data.map((line: unknown[]) => {
             return line.map((value: unknown) => {
-                return `${value}`
+                return toString(value)
             }).join(delimiter)
         }).join(newline)
     }
@@ -43,6 +52,19 @@ export class SquishyIO {
                 return value.trim()
             })
         })
+    }
+
+    xmlParse(content: string, options?: any, mimeType?: string): any {
+        // set default mime type if needed
+        if (!mimeType) {
+            mimeType = SquishyIO.XML_MIME_TYPE
+        }
+        // create the parser
+        const parser: any = new xmldom.DOMParser(options)
+        // parse the document
+        const document: any = parser.parseFromString(content, mimeType)
+
+        return document
     }
 
     numberToString(x: number): string {